Fix build warnings

+++ services/audioflinger/test-resample.cpp
@@ -177,7 +177,7 @@ int main(int argc, char* argv[]) {
             double t = double(i) / input_freq;
             double y = sin(M_PI * k * t * t);
             int16_t yi = floor(y * 32767.0 + 0.5);
-            for (size_t j=0 ; j<channels ; j++) {
+            for (size_t j=0 ; j<(size_t)channels ; j++) {
                 in[i*channels + j] = yi / (1+j);
             }
         }
